Converted MessageID to Long

+++ backend/src/main/java/com/beanstalk/backend/model/MessageDTO.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @Setter
 public class MessageDTO {
 
-    private Integer messageId;
+    private Long messageId;
 
     @NotNull
     @Size(max = 20)
